+++ TentakelsAttacking2/Helper/private/HSoundManager.cpp
@@ -5,7 +5,7 @@
 
 #include "HSoundManager.h"
 #include "UIEvents.hpp"
-#include "Random.h"
+#include "HRandom.h"
 #include <filesystem>
 
 void SoundManager::LoadSounds() {
